Revert the change to ConstructorInvokerTest to use an inner class and move all tests to their own package for consistency
