+++ impl/src/main/java/org/jboss/cdi/tck/tests/invokers/invalid/prodmethod/ProducerMethodBeanInvokerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.invokers.invalid;
+package org.jboss.cdi.tck.tests.invokers.invalid.prodmethod;
 
 import jakarta.enterprise.inject.spi.DeploymentException;
 
@@ -22,28 +22,25 @@ import org.jboss.arquillian.container.test.api.ShouldThrowException;
 import org.jboss.cdi.tck.AbstractTest;
 import org.jboss.cdi.tck.cdi.Sections;
 import org.jboss.cdi.tck.shrinkwrap.WebArchiveBuilder;
-import org.jboss.cdi.tck.tests.invokers.invalid.prodfield.MyProducer;
-import org.jboss.cdi.tck.tests.invokers.invalid.prodfield.MyService;
-import org.jboss.cdi.tck.tests.invokers.invalid.prodfield.TestExtension;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.test.audit.annotations.SpecAssertion;
 import org.jboss.test.audit.annotations.SpecVersion;
 import org.testng.annotations.Test;
 
 @SpecVersion(spec = "cdi", version = "4.1")
-public class ProducerFieldBeanInvokerTest extends AbstractTest {
+public class ProducerMethodBeanInvokerTest extends AbstractTest {
     @Deployment
     @ShouldThrowException(DeploymentException.class)
     public static WebArchive createTestArchive() {
         return new WebArchiveBuilder()
-                .withTestClass(ProducerFieldBeanInvokerTest.class)
+                .withTestClass(ProducerMethodBeanInvokerTest.class)
                 .withClasses(MyService.class, MyProducer.class, TestExtension.class)
                 .withBuildCompatibleExtension(TestExtension.class)
                 .build();
     }
 
     @Test
-    @SpecAssertion(section = Sections.BUILDING_INVOKER, id = "cb")
+    @SpecAssertion(section = Sections.BUILDING_INVOKER, id = "ca")
     public void trigger() {
     }
